+++ picasso/gui/simulate.py
@@ -262,6 +262,13 @@ class Window(QtWidgets.QMainWindow):
         self.setWindowIcon(icon)
         self.initUI()
 
+        self.user_settings_dialog = io.UserSettingsDialog(self)
+        file_menu = self.menuBar().addMenu("File")
+        picasso_settings_action = file_menu.addAction("Picasso settings")
+        picasso_settings_action.triggered.connect(
+            self.user_settings_dialog.show
+        )
+
     def initUI(self):
         self.currentround = CURRENTROUND
         self.structureMode = True